Fix GCC warnings that occur on 64 bit machines.

+++ s3b_config.c
@@ -1028,7 +1028,7 @@ validate_config(void)
     config.num_blocks = config.file_size / config.block_size;
     if (sizeof(s3b_block_t) < sizeof(config.num_blocks)
       && config.num_blocks > ((off_t)1 << (sizeof(s3b_block_t) * 8))) {
-        warnx("more than 2^%d blocks: decrease file size or increase block size", sizeof(s3b_block_t) * 8);
+        warnx("more than 2^%d blocks: decrease file size or increase block size", (int)(sizeof(s3b_block_t) * 8));
         return -1;
     }
